feat: remove duplicate schema
