+++ internal/actions/schema.go
@@ -1,38 +1,22 @@
 package actions
 
 import (
-	"github.com/promiseofcake/artifactsmmo-go-client/client"
-)
+	"time"
 
-// Coords defines X, Y coords as signed integers
-type Coords struct {
-	X int
-	Y int
-}
+	"github.com/promiseofcake/artifactsmmo-go-client/client"
+
+	"github.com/promiseofcake/artifactsmmo-engine/internal/models"
+)
 
 // Response is a generic return value for most actions
 type Response struct {
-	CharacterResponse CharacterResponse
+	CharacterResponse models.Character
 	CooldownSchema    client.CooldownSchema
 }
 
-// GetRemainingCooldown returns the remaining cooldown based upon the last action
-// this is useful for wait loops
-func (g *Response) GetRemainingCooldown() int64 {
-	return int64(g.CooldownSchema.RemainingSeconds)
-}
-
-// CharacterResponse is a specific wrapper around generic return info
-type CharacterResponse struct {
-	client.CharacterSchema
-}
-
-// GetPosition returns the given Coords for the Character in question
-func (c *CharacterResponse) GetPosition() Coords {
-	return Coords{
-		X: c.X,
-		Y: c.Y,
-	}
+// GetCooldownDuration returns the duration for the Response cooldown
+func (g *Response) GetCooldownDuration() time.Duration {
+	return time.Until(g.CooldownSchema.Expiration)
 }
 
 // FightResponse wraps a generic Response with Fight related data
