--- internal/actions/util.go
@@ -1,5 +1,0 @@
-package actions
-
-func toIntPointer(x int) *int {
-	return &x
-}
