+++ internal/cmd/craft.go
@@ -2,12 +2,14 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/spf13/viper"
 	"log/slog"
 	"time"
 
-	"github.com/promiseofcake/artifactsmmo-engine/internal/actions"
+	"github.com/spf13/viper"
+
 	"github.com/spf13/cobra"
+
+	"github.com/promiseofcake/artifactsmmo-engine/internal/actions"
 )
 
 var (
@@ -34,13 +36,13 @@ var craftCmd = &cobra.Command{
 				return fmt.Errorf("failed to craft: %w", err)
 			}
 
-			sec := resp.GetRemainingCooldown()
+			cooldown := resp.GetCooldownDuration()
 			slog.Info("craft results",
 				"xp", resp.SkillInfo.Xp,
 				"items", resp.SkillInfo.Items,
-				"cooldown", sec,
+				"cooldown", cooldown,
 			)
-			time.Sleep(time.Duration(sec) * time.Second)
+			time.Sleep(cooldown)
 		}
 	},
 }
