+++ internal/engine/gather.go
@@ -8,7 +8,6 @@ import (
 	"time"
 
 	"github.com/promiseofcake/artifactsmmo-engine/internal/actions"
-	"github.com/promiseofcake/artifactsmmo-engine/internal/player"
 )
 
 func Gather(ctx context.Context, r *actions.Runner, character string) error {
@@ -35,14 +34,9 @@ func Gather(ctx context.Context, r *actions.Runner, character string) error {
 	slog.Info("we are gathering", "resource", resource)
 
 	// get all character info
-	char, err := r.GetMyCharacterInfo(ctx, character)
+	c, err := r.GetMyCharacterInfo(ctx, character)
 	if err != nil {
 		return fmt.Errorf("failed to get character %w", err)
-	}
-
-	// our mutated state
-	c := player.Character{
-		CharacterSchema: &char.CharacterSchema,
 	}
 
 	// check if we should bank straight away
@@ -58,7 +52,7 @@ func Gather(ctx context.Context, r *actions.Runner, character string) error {
 		}
 		cooldown := time.Until(m.CooldownSchema.Expiration)
 		slog.Info("moved to resource", "resource", resource, "cooldown", cooldown)
-		c.CharacterSchema = &m.CharacterResponse.CharacterSchema
+		c.CharacterSchema = m.CharacterResponse.CharacterSchema
 		time.Sleep(cooldown)
 	}
 
@@ -74,7 +68,7 @@ func Gather(ctx context.Context, r *actions.Runner, character string) error {
 			}
 			cooldown := time.Until(g.CooldownSchema.Expiration)
 			slog.Info("gathered resource", "resource", resource, "result", g.SkillInfo, "cooldown", cooldown)
-			c.CharacterSchema = &g.CharacterResponse.CharacterSchema
+			c.CharacterSchema = g.CharacterResponse.CharacterSchema
 			time.Sleep(cooldown)
 
 			if c.ShouldBank() {
