+++ internal/models/map.go
@@ -1,0 +1,7 @@
+package models
+
+// Coords defines X, Y map coordinates as signed integers
+type Coords struct {
+	X int
+	Y int
+}
